\name{CNES}
\alias{CNES}
\docType{data}
\title{Variables from the 1997 Canadian National Election Study}
\description{
  These variables are from the mailback questionnaire to the 1997 Canadian
  National Election Study, and are intended to tap attitude towards \dQuote{traditional
  values.}
}
\usage{CNES}
\format{
  A data frame with 1529 observations on the following 4 variables.
  \describe{
    \item{\code{MBSA2}}{an ordered factor with levels \code{StronglyDisagree}, 
    \code{Disagree}, \code{Agree}, and \code{StronglyAgree}, in response to the statement,
    \dQuote{We should be more tolerant of people who choose to live according to their own standards, 
    even if they are very different from our own.}}
    
    \item{\code{MBSA7}}{an ordered factor with levels \code{StronglyDisagree}, 
    \code{Disagree}, \code{Agree}, and \code{StronglyAgree}, in response to the statement,
    \dQuote{Newer lifestyles are contributing to the breakdown of our society.}}
    
    \item{\code{MBSA8}}{an ordered factor with levels \code{StronglyDisagree}, 
    \code{Disagree}, \code{Agree}, and \code{StronglyAgree}, in response to the statement,
    \dQuote{The world is always changing and we should adapt our view of moral behaviour to
    these changes.}}
    
    \item{\code{MBSA9}}{an ordered factor with levels \code{StronglyDisagree}, 
    \code{Disagree}, \code{Agree}, and \code{StronglyAgree}, in response to the statement,
    \dQuote{This country would have many fewer problems if there were more emphasis on
    traditional family values.}}
  }
}


\source{
  York University Institute for Social Research.
}

\keyword{datasets}
